Refactoring of the Sidebar component and scss
