+++ src/App.tsx
@@ -1,6 +1,7 @@
 import React, { useCallback, useEffect } from 'react'
 import 'antd/dist/antd.min.css'
 import './App.scss'
+import './pages/components/header/sidebar/sidebar.scss'
 import TimelinePage from 'src/pages/TimelinePage'
 import { ConfigProvider, Layout } from 'antd'
 import 'leaflet/dist/leaflet.css'
@@ -13,7 +14,7 @@ import { PageSearchState, SearchContext } from './model/pageState'
 import moment from 'moment'
 import DashboardPage from './pages/dashboard/DashboardPage'
 import { useSessionStorage } from 'usehooks-ts'
-import SideBar from './pages/components/header/SideBar'
+import SideBar from './pages/components/header/sidebar/SideBar'
 import RealtimeMapPage from './pages/RealtimeMapPage'
 import SingleLineMapPage from './pages/SingleLineMapPage'
 import { useLocation } from 'react-router-dom'
